another change from float to double

+++ Kapitel_08/src/uebung_8_4/Job.java
@@ -2,14 +2,14 @@ package uebung_8_4;
 
 public class Job {
 	private String name;
-	private float salary;
+	private double salary;
 
 	public Job(String name, float salary) {
 		this.name = name;
 		this.salary = salary;
 	}
 
-	public float getSalary() {
+	public double getSalary() {
 		return this.salary;
 	}
